+++ examples/cancel-order.ts
@@ -1,4 +1,4 @@
-import { Flex, Trader } from "../flex";
+import { cancelOrderFinalized, Flex, Trader } from "../flex";
 import { CONFIG, EXAMPLES_FLEX_CONFIG } from "./examples";
 //import { LogLevel } from "../contracts/helpers";
 
@@ -10,24 +10,24 @@ import { CONFIG, EXAMPLES_FLEX_CONFIG } from "./examples";
         const traderId = CONFIG.trader.id;
         const marketAddress = CONFIG.market;
 
-        let orderInfo = await Trader.cancelOrder(
-            flex,
-            {
-                clientAddress: clientAddress,
-                trader: {
-                    id: traderId,
-                    signer: "traderSigner",
-                },
-                marketAddress: marketAddress,
-                price: { tokens: 10 },
-                orderId: "0x1355df445d27aca1",
-                // waitForOrderbookUpdate: true
+        let result = await Trader.cancelOrder(flex, {
+            clientAddress: clientAddress,
+            trader: {
+                id: traderId,
+                signer: "traderSigner",
             },
-        );
+            marketAddress: marketAddress,
+            price: { tokens: 0.2 },
+            orderId: "0xb1482121e43efae",
+            // waitForOrderbookUpdate: true
+        });
 
-        flex.evr.log.info("Order info", orderInfo);
+        flex.evr.log.info("Cancel Initialization result on wallet", result);
 
-
+        if (!cancelOrderFinalized(result)) {
+            result = await Trader.waitForCancelOrder(flex, result);
+            flex.evr.log.info("Finalized cancel result in orderbook", result);
+        }
         await flex.close();
     } catch (err) {
         flex.evr.log.error(err);