+++ examples/make-order.ts
@@ -1,5 +1,6 @@
-import { Flex, Trader } from "../flex";
+import { Flex, makeOrderFinalized, Trader } from "../flex";
 import { CONFIG, EXAMPLES_FLEX_CONFIG } from "./examples";
+import { waitForMakeOrder } from "../flex/trader/make-order";
 //import { LogLevel } from "../contracts/helpers";
 
 (async () => {
@@ -10,7 +11,7 @@ import { CONFIG, EXAMPLES_FLEX_CONFIG } from "./examples";
         const traderId = CONFIG.trader.id;
         const marketAddress = CONFIG.market;
 
-        let orderInfo = await Trader.makeOrder(flex, {
+        let result = await Trader.makeOrder(flex, {
             clientAddress: clientAddress,
             trader: {
                 id: traderId,
@@ -18,15 +19,19 @@ import { CONFIG, EXAMPLES_FLEX_CONFIG } from "./examples";
             },
             sell: false,
             marketAddress: marketAddress,
-            price: { tokens: 10 },
-            amount: { tokens: 2 },
+            price: { tokens: 20 },
+            amount: { tokens: 10 },
             waitForOrderbookUpdate: true,
+            
         });
-        
-        flex.evr.log.info("Order info", orderInfo);
+        flex.evr.log.info("MakeOrder Initialization result on wallet", result);
+
+        if (!makeOrderFinalized(result)) {
+            result = await waitForMakeOrder(flex, result);
+            flex.evr.log.info("Finalized Make order result in orderbook", result);
+        }
 
         await flex.close();
-
     } catch (err) {
         flex.evr.log.error(err);
         process.exit(1);