+++ flex/trader/index.ts
@@ -2,8 +2,20 @@
  * @module trader
  */
 import { Flex } from "../flex";
-import { makeOrder, MakeOrderOptions, NewOrderInfo } from "./make-order";
-import { cancelOrder, CancelOrderOptions, CancelOrderResult } from "./cancel-order";
+import {
+    makeOrder,
+    MakeOrderOptions,
+    MakeOrderResult,
+    MakeOrderStatus,
+    waitForMakeOrder,
+} from "./make-order";
+import {
+    cancelOrder,
+    CancelOrderOptions,
+    CancelOrderResult,
+    CancelOrderStatus,
+    waitForCancelOrder,
+} from "./cancel-order";
 import { OrderInfo, TradeInfo } from "./types";
 import { queryOrders, queryTrades, queryWallets, QueryWalletsOptions } from "./query";
 import { WalletInfo } from "../client";
@@ -15,24 +27,26 @@ import {
 } from "./deploy-ever-wallet";
 import {
     deployTraderTip31Wallet,
-    DeployTraderTip31WalletOptions, DeployTraderTip31WalletResult,
+    DeployTraderTip31WalletOptions,
+    DeployTraderTip31WalletResult,
 } from "./deploy-tip31-wallet";
 
 export * from "./types";
 export {
     DeployTraderOptions,
     MakeOrderOptions,
-    NewOrderInfo,
+    MakeOrderResult,
+    MakeOrderStatus,
+    CancelOrderStatus,
     CancelOrderOptions,
+    CancelOrderResult,
     QueryWalletsOptions,
     DeployTraderTip31WalletOptions,
     DeployTraderEverWalletOptions,
     WalletInfo,
     EverWalletInfo,
     DeployTraderTip31WalletResult,
-    CancelOrderResult,
 };
-
 
 export class Trader {
     /**
@@ -82,30 +96,88 @@ export class Trader {
         return await deployTraderTip31Wallet(flex, options);
     }
 
-   /**
-    * Places an order on a specified market
-    * @param flex
-    * DEX instance
-    * @param options
-    * Order options
-    * @returns
-    */
-    static async makeOrder(flex: Flex, options: MakeOrderOptions): Promise<NewOrderInfo> {
+    /**
+     * Places an order on a specified market.
+     *
+     * If this function throws error then making order message was not delivered to the blockchain.
+     * Otherwise, this function returns result that depends on `status` field.
+     * User can resume finalizing with `waitForMakeOrder` if result is not finalized yet
+     * (neither success nor error).
+     *
+     * @param flex
+     * DEX instance
+     *
+     * @param options
+     * Order options
+     *
+     * @returns
+     */
+    static async makeOrder(flex: Flex, options: MakeOrderOptions): Promise<MakeOrderResult> {
         return await makeOrder(flex, options);
     }
 
     /**
-     * Cancels an Order on DEX Market
+     * Resumes previously terminated makeOrder.
+     *
+     * If this function throws error then waiting process can not be done at this time.
+     * Otherwise, this function returns new result that depends on `status` field.
+     * User can resume finalizing with `waitForMakeOrder` if result is not finalized yet
+     * (neither success nor error).
      *
      * @param flex
      * DEX instance
-     * @param options
-     * Cancel options
+     *
+     * @param result
+     * Order processing result.
+     *
      * @returns
      */
+    static async waitForMakeOrder(flex: Flex, result: MakeOrderResult): Promise<MakeOrderResult> {
+        return await waitForMakeOrder(flex, result);
+    }
+
+    /**
+     * Cancels an order.
+     *
+     * If this function throws error then cancellation message was not delivered to the blockchain.
+     * Otherwise, this function returns result that depends on `status` field.
+     * User can resume finalizing with `waitForCancelOrder` if result is not finalized yet
+     * (neither success nor error).
+     *
+     * @param flex
+     * DEX instance
+     *
+     * @param options
+     * Cancellation options
+     *
+     * @returns     */
     static async cancelOrder(flex: Flex, options: CancelOrderOptions): Promise<CancelOrderResult> {
         return await cancelOrder(flex.evr, options);
     }
+
+    /**
+     * Resumes previously terminated `cancelOrder`.
+     *
+     * If this function throws error then waiting process can not be done at this time.
+     * Otherwise, this function returns new result that depends on `status` field.
+     * User can resume finalizing with `waitForCancelOrder` if result is not finalized yet
+     * (neither success nor error).
+     *
+     * @param flex
+     * DEX instance
+     *
+     * @param result
+     * Order cancellation result.
+     *
+     * @returns
+     */
+    static async waitForCancelOrder(
+        flex: Flex,
+        result: CancelOrderResult,
+    ): Promise<CancelOrderResult> {
+        return await waitForCancelOrder(flex.evr, result);
+    }
+
     /**
      * Returns Trader's open orders on DEX
      * @param flex
@@ -132,10 +204,15 @@ export class Trader {
      * @param options
      * @returns
      */
-    static async queryWallets(
-        flex: Flex,
-        options: QueryWalletsOptions,
-    ): Promise<WalletInfo[]> {
+    static async queryWallets(flex: Flex, options: QueryWalletsOptions): Promise<WalletInfo[]> {
         return await queryWallets(flex, options);
     }
+}
+
+export function makeOrderFinalized(result: MakeOrderResult): boolean {
+    return result.status === MakeOrderStatus.SUCCESS || result.status === MakeOrderStatus.ERROR;
+}
+
+export function cancelOrderFinalized(result: CancelOrderResult): boolean {
+    return result.status === CancelOrderStatus.SUCCESS || result.status === CancelOrderStatus.ERROR;
 }